move icons and components to their own sections in `$lib`
