+++ src/lib/components/SelectPlace/index.ts
@@ -1,0 +1,2 @@
+import SelectPlace from "./SelectPlace.svelte";
+export { SelectPlace };
