+++ src/lib/components/index.ts
@@ -1,0 +1,13 @@
+import Combobox from "./Combobox";
+export { Combobox };
+
+import InsertEntry from "./Header/InsertEntry.svelte";
+export { InsertEntry };
+
+export * from "./SelectPlace";
+
+import TimezoneTable from "./TimezoneTable/Table.svelte";
+export { TimezoneTable };
+
+import UserPreferences from "./Header/UserPreferences.svelte";
+export { UserPreferences };
