+++ src/lib/icons/DeleteIcon.svelte
