+++ src/lib/icons/InfoIcon.svelte
