+++ src/lib/icons/InsertIcon.svelte
