+++ src/lib/icons/SearchIcon.svelte
