+++ src/lib/icons/SettingsIcon.svelte
