+++ src/lib/icons/index.ts
@@ -1,0 +1,17 @@
+import DeleteIcon from "./DeleteIcon.svelte";
+export { DeleteIcon };
+
+import EditIcon from "./EditIcon.svelte";
+export { EditIcon };
+
+import InfoIcon from "./InfoIcon.svelte";
+export { InfoIcon };
+
+import InsertIcon from "./InsertIcon.svelte";
+export { InsertIcon };
+
+import SearchIcon from "./SearchIcon.svelte";
+export { SearchIcon };
+
+import SettingsIcon from "./SettingsIcon.svelte";
+export { SettingsIcon };
