--- src/lib/index.ts
@@ -1,11 +1,0 @@
-import Combobox from "./components/Combobox";
-export { Combobox };
-
-import InsertEntry from "./components/Header/InsertEntry.svelte";
-export { InsertEntry };
-
-import TimezoneTable from "./components/TimezoneTable/Table.svelte";
-export { TimezoneTable };
-
-import UserPreferences from "./components/Header/UserPreferences.svelte";
-export { UserPreferences };
